tail: Implement `-F` and `--retry` options
Fixes #23097

The retry option will keep trying to open the file in read mode
until it is successfully accessible. Thenafter, if the file becomes
inaccessible, no error will be reported. As a result, this will work
during the initial check.

The -F is a shorthand to set both --follow and --retry flags.